requirements.txt
backports.csv==1.0.2
click==6.6
colorama==0.3.7
cycler==0.10.0
matplotlib==1.5.3
nose==1.3.7
numpy==1.11.2
Pillow==3.4.1
pyparsing==2.1.10
python-dateutil==2.5.3
python-termstyle==0.1.10
pytz==2016.7
rednose==1.2.1
scipy==0.18.1
six==1.10.0
wordcloud==1.2.1
